Split out helper project for Wasm, as described in #18
