+++ src/Blazor.FileReader.Wasm.Demo/Startup.cs
@@ -9,27 +9,7 @@ namespace Blazor.FileReader.Wasm.Demo
     {
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddSingleton<IInvokeUnmarshalled, WasmJsRuntimeInvokeProvider>();
-            services.AddSingleton<IFileReaderService, FileReaderService>();
-        }
-
-        /// <summary>
-        /// Provides an optional bridge for unmarshalled calls
-        /// </summary>
-        private class WasmJsRuntimeInvokeProvider : IInvokeUnmarshalled
-        {
-            public IJSRuntime CurrentJSRuntime { get; }
-
-            public WasmJsRuntimeInvokeProvider(IJSRuntime currentJSRuntime)
-            {
-                CurrentJSRuntime = currentJSRuntime;
-            }
-
-            private MonoWebAssemblyJSRuntime MonoWebAssemblyJSRuntime =>
-                CurrentJSRuntime as MonoWebAssemblyJSRuntime;
-
-            public TRes InvokeUnmarshalled<T1, T2, TRes>(string identifier, T1 arg1, T2 arg2) =>
-                MonoWebAssemblyJSRuntime.InvokeUnmarshalled<T1, T2, TRes>(identifier, arg1, arg2);
+            services.AddFileReader();
         }
 
         public void Configure(IComponentsApplicationBuilder app)
