+++ src/Blazor.FileReader.Wasm/Blazor.FileReader.Wasm.csproj
@@ -1,0 +1,26 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <Version>0.9.0-preview-13032019</Version>
+    <PackageId>Tewr.Blazor.FileReader.Wasm</PackageId>
+    <Authors>Tor Knutsson</Authors>
+    <Description>Small helper library that helps setting up BlazorFileReader (</Description>
+    <PackageProjectUrl>https://github.com/Tewr/BlazorFileReader</PackageProjectUrl>
+    <RepositoryUrl>https://github.com/Tewr/BlazorFileReader</RepositoryUrl>
+    <PackageTags>blazor blazor-component stream filestream file-stream read-file</PackageTags>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <PackageLicenseFile>LICENSE</PackageLicenseFile>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.AspNetCore.Blazor" Version="0.9.0-preview3-19154-02" />
+    <PackageReference Include="Microsoft.AspNetCore.Blazor.Build" Version="0.9.0-preview3-19154-02" PrivateAssets="all" />
+    <PackageReference Include="Tewr.Blazor.FileReader" Version="0.9.0-preview-13032019" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="..\..\LICENSE">
+      <Pack>True</Pack>
+      <PackagePath></PackagePath>
+    </None>
+  </ItemGroup>
+</Project>
